Updated to version 0.09.03 Fixed compression check bug. Modified default compression check size. Fixed compiler warnings for Windows x64.

+++ src/maidsafe/encrypt/utils.cc
@@ -56,7 +56,7 @@ bool CheckCompressibility(const std::string &sample,
     return false;
 
   std::string compressed_sample(Compress(sample, self_encryption_type));
-  double ratio = compressed_sample.size() / sample.size();
+  double ratio = static_cast<double>(compressed_sample.size()) / sample.size();
   return !compressed_sample.empty() && ratio <= 0.9;
 }
